feat: add participation status to vote options and implement recommendation algorithm
