+++ backend/internal/models/models.go
@@ -121,10 +121,11 @@ type VoteAnswer struct {
 
 type VoteOption struct {
 	Model
-	Vote        Vote         `gorm:"foreignKey:VoteID;references:ID;constraint:OnDelete:RESTRICT" json:"vote"`
-	VoteID      uint         `json:"voteId"`
-	Text        string       `gorm:"not null;size:255" json:"text"`
-	VoteAnswers []VoteAnswer `gorm:"foreignKey:VoteOptionID" json:"voteAnswers"`
+	Vote                Vote                `gorm:"foreignKey:VoteID;references:ID;constraint:OnDelete:RESTRICT" json:"vote"`
+	VoteID              uint                `json:"voteId"`
+	Text                string              `gorm:"not null;size:255" json:"text"`
+	ParticipationStatus ParticipationStatus `gorm:"not null" json:"participationStatus"`
+	VoteAnswers         []VoteAnswer        `gorm:"foreignKey:VoteOptionID" json:"voteAnswers"`
 }
 
 type Opinion struct {
